账号工具类的封装
保存账号信息:
         数据存储一般会搞一个业务类,专门处理数据
         以后不想归档,用数据库,直接修改业务类

+++ Weibo/Weibo/HJStatusFrame.m
@@ -102,7 +102,7 @@
     if (_status.pic_urls.count) { // 有配图
         CGFloat photosX = HJCellMargin;
         CGFloat photosY = originH;
-        CGSize photosSize = [HJPhotosView photosSizeWithCount:_status.pic_urls.count];
+        CGSize photosSize = [HJPhotosView photosSizeWithCount:(int)_status.pic_urls.count];
         _photosViewF = (CGRect){{photosX,photosY},photosSize};
         
         originH = CGRectGetMaxY(_photosViewF) + HJCellMargin;
